refactor: 정수형으로 반환하도록 타입 수정

+++ src/models/schemas.py
@@ -33,6 +33,6 @@ class AIAnalysisResponse(BaseModel):
     isWarning: bool
     keywords: List[str]
     sentimentComments: List[CommentSentimentDetail]
-    languageRatio: Dict[str, float]
-    sentimentRatio: Dict[str, float]
+    languageRatio: Dict[str, int]
+    sentimentRatio: Dict[str, int]
